+++ src/crawlee/crawlers/_beautifulsoup/_beautifulsoup_crawler.py
@@ -18,7 +18,7 @@ if TYPE_CHECKING:
     from crawlee.crawlers._abstract_http import ParsedHttpCrawlingContext
 
 
-@docs_group('Classes')
+@docs_group('Crawlers')
 class BeautifulSoupCrawler(AbstractHttpCrawler[BeautifulSoupCrawlingContext, BeautifulSoup, Tag]):
     """A web crawler for performing HTTP requests and parsing HTML/XML content.