+++ src/crawlee/crawlers/_parsel/_parsel_parser.py
@@ -6,6 +6,7 @@ from typing import TYPE_CHECKING
 from parsel import Selector
 from typing_extensions import override
 
+from crawlee._utils.docs import docs_group
 from crawlee.crawlers._abstract_http import AbstractHttpParser
 
 if TYPE_CHECKING:
@@ -14,6 +15,7 @@ if TYPE_CHECKING:
     from crawlee.http_clients import HttpResponse
 
 
+@docs_group('HTTP parsers')
 class ParselParser(AbstractHttpParser[Selector, Selector]):
     """Parser for parsing HTTP response using Parsel."""