+++ src/crawlee/fingerprint_suite/_fingerprint_generator.py
@@ -3,13 +3,10 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING
 
-from crawlee._utils.docs import docs_group
-
 if TYPE_CHECKING:
     from browserforge.fingerprints import Fingerprint
 
 
-@docs_group('Abstract classes')
 class FingerprintGenerator(ABC):
     """A class for creating browser fingerprints that mimic browser fingerprints of real users."""