+++ src/crawlee/sessions/_cookies.py
@@ -12,7 +12,7 @@ if TYPE_CHECKING:
     from collections.abc import Iterator
 
 
-@docs_group('Data structures')
+@docs_group('Session management')
 class CookieParam(TypedDict, total=False):
     """Dictionary representation of cookies for `SessionCookies.set` method."""
 
@@ -52,9 +52,10 @@ class PlaywrightCookieParam(TypedDict, total=False):
     httpOnly: NotRequired[bool]
     expires: NotRequired[float]
     sameSite: NotRequired[Literal['Lax', 'None', 'Strict']]
+    partitionKey: NotRequired[str | None]
 
 
-@docs_group('Data structures')
+@docs_group('Session management')
 class SessionCookies:
     """Storage cookies for session with browser-compatible serialization and deserialization."""