+++ src/crawlee/storage_clients/_file_system/_request_queue_client.py
@@ -246,7 +246,6 @@ class FileSystemRequestQueueClient(RequestQueueClient):
                     had_multiple_clients=False,
                     handled_request_count=0,
                     pending_request_count=0,
-                    stats={},
                     total_request_count=0,
                 )
                 client = cls(