+++ src/crawlee/storage_clients/_memory/_request_queue_client.py
@@ -93,7 +93,6 @@ class MemoryRequestQueueClient(RequestQueueClient):
             had_multiple_clients=False,
             handled_request_count=0,
             pending_request_count=0,
-            stats={},
             total_request_count=0,
         )