+++ src/crawlee/storage_clients/models.py
@@ -13,7 +13,7 @@ from crawlee._utils.urls import validate_http_url
 KvsValueType = TypeVar('KvsValueType', default=Any)
 
 
-@docs_group('Data structures')
+@docs_group('Storage data')
 class StorageMetadata(BaseModel):
     """Represents the base model for storage metadata.
 
@@ -38,7 +38,7 @@ class StorageMetadata(BaseModel):
     """The timestamp when the storage was last modified."""
 
 
-@docs_group('Data structures')
+@docs_group('Storage data')
 class DatasetMetadata(StorageMetadata):
     """Model for a dataset metadata."""
 
@@ -48,14 +48,14 @@ class DatasetMetadata(StorageMetadata):
     """The number of items in the dataset."""
 
 
-@docs_group('Data structures')
+@docs_group('Storage data')
 class KeyValueStoreMetadata(StorageMetadata):
     """Model for a key-value store metadata."""
 
     model_config = ConfigDict(populate_by_name=True, from_attributes=True)
 
 
-@docs_group('Data structures')
+@docs_group('Storage data')
 class RequestQueueMetadata(StorageMetadata):
     """Model for a request queue metadata."""
 
@@ -70,14 +70,11 @@ class RequestQueueMetadata(StorageMetadata):
     pending_request_count: Annotated[int, Field(alias='pendingRequestCount')]
     """The number of requests that are still pending in the queue."""
 
-    stats: Annotated[dict, Field(alias='stats')]
-    """Statistics about the request queue, TODO?"""
-
     total_request_count: Annotated[int, Field(alias='totalRequestCount')]
     """The total number of requests that have been added to the queue."""
 
 
-@docs_group('Data structures')
+@docs_group('Storage data')
 class KeyValueStoreRecordMetadata(BaseModel):
     """Model for a key-value store record metadata."""
 
@@ -99,7 +96,7 @@ class KeyValueStoreRecordMetadata(BaseModel):
     """The size of the record in bytes."""
 
 
-@docs_group('Data structures')
+@docs_group('Storage data')
 class KeyValueStoreRecord(KeyValueStoreRecordMetadata, Generic[KvsValueType]):
     """Model for a key-value store record."""
 
@@ -109,7 +106,7 @@ class KeyValueStoreRecord(KeyValueStoreRecordMetadata, Generic[KvsValueType]):
     """The value of the record."""
 
 
-@docs_group('Data structures')
+@docs_group('Storage data')
 class DatasetItemsListPage(BaseModel):
     """Model for a single page of dataset items returned from a collection list method."""
 
@@ -134,7 +131,7 @@ class DatasetItemsListPage(BaseModel):
     """The list of dataset items returned on this page."""
 
 
-@docs_group('Data structures')
+@docs_group('Storage data')
 class ProcessedRequest(BaseModel):
     """Represents a processed request."""
 
@@ -146,7 +143,7 @@ class ProcessedRequest(BaseModel):
     was_already_handled: Annotated[bool, Field(alias='wasAlreadyHandled')]
 
 
-@docs_group('Data structures')
+@docs_group('Storage data')
 class UnprocessedRequest(BaseModel):
     """Represents an unprocessed request."""
 
@@ -157,7 +154,7 @@ class UnprocessedRequest(BaseModel):
     method: Annotated[HttpMethod | None, Field()] = None
 
 
-@docs_group('Data structures')
+@docs_group('Storage data')
 class AddRequestsResponse(BaseModel):
     """Model for a response to add requests to a queue.